refact: move serialization into dedicated `mod serialization`
